Вынес в отдельные файлы
