+++ Dictionary/Exceptions/RecordAlreadyExistException.cs
@@ -1,0 +1,6 @@
+﻿namespace Dictionary.Exceptions;
+
+public class RecordAlreadyExistException : Exception
+{
+    public RecordAlreadyExistException( string message ) : base( message ) { }
+}
