+++ Dictionary/Exceptions/RocordNotFoundException.cs
@@ -1,0 +1,6 @@
+﻿namespace Dictionary.Exceptions;
+
+public class RecordNotFoundException : Exception
+{
+    public RecordNotFoundException( string message ) : base( message ) { }
+}
